app/pip-requirements.txt
requests==2.25.1 
django==3.2.4
python-telegram-bot==13.5
psycopg2-binary>=2.8